requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-filter==24.2
psycopg2-binary==2.9.9
django-phonenumber-field==7.3.0
djangorestframework-simplejwt==5.3.1
phonenumbers==8.13.40